fix(cli/tui): review fixes for create form (#1748)
- Move validate_env_entry from mount.py to new env.py (wrong module)
- Replace bare except Exception with specific types + debug logging
  in action_create (httpx.HTTPError, OSError, ValueError)
- Narrow JSON-parsing fallback to except (ValueError, KeyError)
- Add explicit Select.BLANK/Select.NULL guard in _create for clarity
